Add new HTML index page, Github workflows, and Dockerfiles
This commit includes a new homepage for the ExampleApp, stored in WWWroot/index.html. This page contains a simple client-side UI and instructions on how to use the JsonServiceClient in web pages. This commit also introduces a new Github release workflow, ".github/workflows/release.yml", which comprises steps to build and push docker images using Github actions and deploy via SSH. The addition of a new Docker Compose file included in the .deploy directory completes this dockerization process. Finally, this update creates an auto-generated Types